requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
scikit-learn>=1.3.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
ipykernel
kaggle
seaborn
streamlit